src/backend/requirements.txt
Flask==3.1.0
openai==1.59.5
python-dotenv==1.0.1
